feaf: useDiscountCalculator 훅 제거, util함수로 분리
