+++ src/refactoring/hooks/index.ts
@@ -1,4 +1,5 @@
 export * from "./useCart.ts";
+export * from "./useCartStorage.ts";
 export * from "./useCoupon.ts";
 export * from "./useProduct.ts";
 export * from "./useProductSearch.ts";
